# api-definition.yml
openapi: 3.0.3
info:
  title: Mini Doodle Scheduler API
  description: >
    Meeting scheduling platform API. Users manage time slots on their
    personal calendar and convert slots into meetings with participants.
  version: "1.0.0"
servers:
  - url: /
tags:
  - name: Users
    description: User management
  - name: Slots
    description: Time slot management
  - name: Meetings
    description: Meeting scheduling
paths:
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestDto'
      responses:
        '201':
          description: User created
          headers:
            Location:
              description: URL of the created user
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Validation failed
  /api/v1/users/{userId}:
    get:
      tags: [Users]
      summary: Get a user by id
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: User not found
  /api/v1/users/{userId}/slots:
    post:
      tags: [Slots]
      summary: Create a time slot
      operationId: createSlot
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSlotRequestDto'
      responses:
        '201':
          description: Slot created
          headers:
            Location:
              description: URL of the created slot
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotResponseDto'
        '400':
          description: Validation failed
        '404':
          description: User not found
    get:
      tags: [Slots]
      summary: List time slots for a user
      operationId: listSlots
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SlotStatusDto'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Page of slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotPageDto'
        '404':
          description: User not found
  /api/v1/users/{userId}/slots/{slotId}:
    patch:
      tags: [Slots]
      summary: Modify a time slot
      operationId: updateSlot
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSlotRequestDto'
      responses:
        '200':
          description: Slot updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotResponseDto'
        '400':
          description: Validation failed
        '404':
          description: User or slot not found
        '409':
          description: Slot is booked and cannot be modified
    delete:
      tags: [Slots]
      summary: Delete a time slot
      operationId: deleteSlot
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Slot deleted
        '404':
          description: User or slot not found
        '409':
          description: Slot is booked and cannot be deleted
  /api/v1/users/{userId}/slots/{slotId}/meetings:
    post:
      tags: [Meetings]
      summary: Book a slot into a meeting
      operationId: bookSlot
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeetingRequestDto'
      responses:
        '201':
          description: Meeting created
          headers:
            Location:
              description: URL of the created meeting
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponseDto'
        '400':
          description: Validation failed, or a participant does not exist
        '404':
          description: User or slot not found
        '409':
          description: Slot is already booked
  /api/v1/meetings/{meetingId}:
    get:
      tags: [Meetings]
      summary: Get a meeting by id
      operationId: getMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponseDto'
        '404':
          description: Meeting not found
    patch:
      tags: [Meetings]
      summary: Modify a meeting's title, description, and participants
      operationId: updateMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeetingRequestDto'
      responses:
        '200':
          description: Meeting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponseDto'
        '400':
          description: Validation failed, or a participant does not exist
        '404':
          description: Meeting not found
        '409':
          description: Meeting is cancelled and cannot be modified
    delete:
      tags: [Meetings]
      summary: Cancel a meeting
      operationId: cancelMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Meeting cancelled, slot reverted to free
        '404':
          description: Meeting not found
        '409':
          description: Meeting is already cancelled
  /api/v1/users/{userId}/meetings:
    get:
      tags: [Meetings]
      summary: List a user's meetings
      operationId: listMeetings
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MeetingRoleDto'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Page of meetings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingPageDto'
        '404':
          description: User not found
components:
  schemas:
    CreateUserRequestDto:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
    UserResponseDto:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    SlotStatusDto:
      type: string
      enum: [FREE, BUSY]
    CreateSlotRequestDto:
      type: object
      required: [startTime, durationMinutes]
      properties:
        startTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          minimum: 1
    SlotResponseDto:
      type: object
      required: [id, userId, startTime, endTime, status]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SlotStatusDto'
    SlotPageDto:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SlotResponseDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
    MeetingStatusDto:
      type: string
      enum: [SCHEDULED, CANCELLED]
    CreateMeetingRequestDto:
      type: object
      required: [title, participantUserIds]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        participantUserIds:
          type: array
          items:
            type: string
            format: uuid
    MeetingResponseDto:
      type: object
      required: [id, slotId, organizerId, title, status, participantUserIds, createdAt]
      properties:
        id:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        organizerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/MeetingStatusDto'
        participantUserIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
    MeetingRoleDto:
      type: string
      enum: [ORGANIZER, PARTICIPANT, ANY]
      default: ANY
    MeetingPageDto:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MeetingResponseDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64